backend/django/django_angular/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
